requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
numpy==1.24.3
chromadb==0.4.24
groq==0.4.1
httpx==0.26.0
pydantic==2.6.3
sqlalchemy==2.0.27
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
aiohttp
python-dotenv 
pydantic-settings 
aiohttp